fix(news): render news-alert footer without the client-only identity hook
#67 removed the EmailIdentityProvider element, but the send then failed at
"Attempted to call useEmailIdentity() from the server" — Footer still invoked
the `'use client'` context hook, which an App Router server route can't call.

Extract a pure, prop-only FooterContent (no hook, default identity lives in
this server-safe module). NewsAlert renders FooterContent directly with the
`identity` prop, so the news-alert render tree references nothing from the
`'use client'` email-identity module. The hook-based Footer is kept as a thin
wrapper over FooterContent for every other template (rendered via
get-email-content), so their context-driven branding is unchanged.
