+++ apps/email-builder/components/email-identity.tsx
@@ -2,17 +2,12 @@
 
 import React, { createContext, useContext } from 'react'
 import type { EmailIdentity } from '@my/app/types/brand-profile'
+// Default lives in the server-safe FooterContent module (single source of
+// truth) so it can also be used from App Router server routes; re-exported here
+// for existing importers.
+import { DEFAULT_EMAIL_IDENTITY } from './FooterContent'
 
-/**
- * Brand identity for the email footer, supplied at render time. Defaults to
- * Toronto East so any template rendered without a provider is unchanged
- * (no regression). Callers wrap the rendered element with
- * <EmailIdentityProvider value={...}> to brand the footer per tenant/org.
- */
-export const DEFAULT_EMAIL_IDENTITY: EmailIdentity = {
-  name: 'Toronto East Christadelphians',
-  addressLines: ['975 Cosburn Avenue', 'Toronto, On M4C 2W8', 'Canada'],
-}
+export { DEFAULT_EMAIL_IDENTITY }
 
 const EmailIdentityContext = createContext<EmailIdentity>(DEFAULT_EMAIL_IDENTITY)
